모임 생성 로직 고도화. - 모임 생성시 활동 지역과 모임 관심사를 지정하여 생성하도록 만듦 - 모임 생성시 생성하는 유저에게 모임장 권한 부여

+++ src/main/kotlin/com/taskforce/superinvention/app/domain/state/ClubState.kt
@@ -11,5 +11,5 @@ class ClubState(
         var club:Club,
         @ManyToOne
         var state: State,
-        var priority: Int
+        var priority: Long
 ) : BaseEntity()
